1.6 updates
I'm not putting every change here, read the mod page changelogs

+++ Quality Bait/Config.cs
@@ -1,9 +1,4 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace QualityBait
 {
@@ -12,6 +7,6 @@ namespace QualityBait
         public int ChancePercentage { get; set; } = 75;
 
         [JsonIgnore]
-        public double Chance => (double)ChancePercentage / 100.0;
+        public double Chance => ChancePercentage / 100.0;
     }
 }
